Add a PHP 8 attribute for configuration
