+++ composer.json
@@ -20,11 +20,12 @@
 
     "require": {
         "php": "^7.1|8.0.*|8.1.*",
+        "doctrine/annotations": "^1.0",
         "symfony/config": "^4.4 | ^5.0 | ^6.0",
         "symfony/dependency-injection": "^4.4 | ^5.0 | ^6.0",
+        "symfony/deprecation-contracts": "^2.0|^3.0",
         "symfony/http-foundation": "^4.4 | ^5.0 | ^6.0",
-        "symfony/http-kernel": "^4.4 | ^5.0 | ^6.0",
-        "doctrine/annotations": "^1.0"
+        "symfony/http-kernel": "^4.4 | ^5.0 | ^6.0"
     },
 
     "require-dev": {
